backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
qdrant-client==1.15.0
openai==1.58.1
PyPDF2==3.0.1
python-multipart==0.0.20
python-dotenv==1.0.1
tiktoken==0.8.0
pydantic==2.10.3
numpy==2.2.1
setuptools==75.8.0
